Refactors test-bb to use core
