--- impl/mono/core/Misc/Class1.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace sb.Core.Misc
-{
-    public class Class1
-    {
-        public int Test(int a)
-        {
-            return a + 1;
-        }
-    }
-}
