+++ impl/mono/test-bb/Program.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using sb_bbt;
 
 namespace sb.TestBB
 {
