+++ impl/mono/test-bb/Tests/Fetch/TestFetchDeep.cs
@@ -1,9 +1,9 @@
 ﻿using System.IO;
 using System.Linq;
-using sb.Test.Utils;
-using sb_bbt.Utils;
+using sb.Core.Utils;
+using sb.TestBB.Utils;
 
-namespace sb_bbt.Tests.Fetch
+namespace sb.TestBB.Tests.Fetch
 {
     public class TestFetchDeep : Test
     {
@@ -15,7 +15,7 @@ namespace sb_bbt.Tests.Fetch
         public override void Run()
         {
             WriteStart();
-            Fs.DeleteFolder(App.EnvDir);
+            FsUtils.DeleteFolder(App.EnvDir);
             Fs.RunProcess("sb", "fetch", Output);
             var dirCount = new DirectoryInfo(App.EnvDir).GetDirectories().Count();
             var fileCount = new DirectoryInfo(App.EnvDir).GetFiles().Count();
